third_party/WebKit/LayoutTests/fast/dom/shadow/event-path-multiple-shadow-roots.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
</head>
<body>
<p id="description"></p>
<div id="sandbox"></div>
<pre id="console"></pre>
<script>
var sandbox = document.getElementById('sandbox');

function click(path) {
    debug('');
    debug('click #' + path);
    var clickEvent = document.createEvent("MouseEvents");
    clickEvent.initMouseEvent('click', true, false, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
    getNodeInComposedTree(path).dispatchEvent(clickEvent);
}

sandbox.appendChild(
    createDOM('div', {'id': 'A'},
              createDOM('div', {'id': 'B'},
                        createDOM('div', {'id': 'C'})),
              createShadowRoot({'id': 'D'},
                               createDOM('div', {'id': 'E'},
                                         createDOM('content', {'id': 'F'}))),
              createShadowRoot({'id': 'G'},
                               createDOM('div', {'id': 'H'},
                                         createDOM('shadow', {'id': 'I'})))));

['A', 'B', 'C', 'A/', 'A/E', 'A/F', 'A//', 'A//H', 'A//I'].forEach(function(path) {
    getNodeInComposedTree(path).addEventListener('click', function(event) {
        debug('\nevent.path on node ' + dumpNode(event.currentTarget) + ' (target: ' + dumpNode(event.target) + ')');
        debug(dumpNodeList(event.path));
    });
});

click('C');
click('A/E');
click('A//H');
</script>
</body>
</html>
